Partially working GL render

+++ src/platform.h
@@ -8,7 +8,7 @@ vec2i_t platform_screen_size(void);
 double platform_now(void);
 bool platform_get_fullscreen(void);
 void platform_set_fullscreen(bool fullscreen);
-void platform_set_audio_mix_cb(void (*cb)(float *buffer, uint32_t len));
+void platform_set_audio_mix_cb(void (*cb)(uint32_t *buffer, uint32_t len));
 
 uint8_t *platform_load_asset(const char *name, uint32_t *bytes_read);
 uint8_t *platform_load_userdata(const char *name, uint32_t *bytes_read);
